+++ src/core/math/vector4.h
@@ -146,7 +146,7 @@ inline float dot(const Vector4& a, const Vector4& b)
 /// Returns the lenght of @a a.
 inline float length(const Vector4& a)
 {
-	return sqrt(a.x * a.x + a.y * a.y + a.z * a.z + a.w * a.w);
+	return sqrtf(a.x * a.x + a.y * a.y + a.z * a.z + a.w * a.w);
 }
 
 /// Returns the squared length of @a a.
